Replace BankIdLoginViewModelTest with SwedishLoginPresenterTest
